mac18

+++ lab-2-4-2/lab-2-4-2.c
@@ -28,9 +28,9 @@ const int FndFont_dot[10] = {0xBF, 0x86, 0xDB, 0xCF, 0xE6, 0xED, 0xFD, 0x87, 0xF
 
 // Set timer
 unsigned long pre_time = 0;
-unsigned long pre_time_2 = 0;
+unsigned double pre_time_2 = 0;
 unsigned long cur_time = 0;
-unsigned long cur_time_2 = 0;
+unsigned double cur_time_2 = 0;
 const int duration = 100;
 const double duration_2 = 10;
 
@@ -98,7 +98,7 @@ int main() {
 
     while (1) {
     	cur_time_2 = millis();
-    	long time_unit = cur_time_2 - pre_time_2;
+    	double time_unit = cur_time_2 - pre_time_2;
     	double time_unit_no_float =floor(time_unit);
 
     	if (time_unit_no_float >= duration_2) {
